+++ src/botInteractions/buttonInteractions/ticket.ts
@@ -11,8 +11,8 @@ import {
 } from "discord.js";
 import { Button } from "../../classes/button.js";
 import { GamerBotAPIInstance } from "../../index.js";
-import NoteCommand from "../commands/admin_commands/note.js";
-import { createButtonRow, createModal } from "../../functions/builder_functions.js";
+import NoteCommand from "../commands/adminCommands/note.js";
+import { createButtonRow, createModal } from "../../functions/builderFunctions.js";
 
 export default class Ticket implements Button {
     name = "ticket";
