+++ src/botInteractions/commands/adminCommands/kick.ts
@@ -1,19 +1,19 @@
 import {
-    ChatInputCommandInteraction,
+    SlashCommandBuilder,
     GuildMember,
     PermissionFlagsBits,
-    SlashCommandBuilder,
+    ChatInputCommandInteraction,
 } from "discord.js";
 import { Command } from "../../../classes/command.js";
 import { GamerBotAPIInstance } from "../../../index.js";
 import { ModLog } from "../../../classes/modlog.js";
-import { CreateModLogEmbed } from "../../../functions/builder_functions.js";
+import { CreateModLogEmbed } from "../../../functions/builderFunctions.js";
 
-export default class NoteCommand implements Command {
-    name = "note";
+export default class KickCommand implements Command {
+    name = "kick";
     ephemeral = false;
     defer = true;
-    description = "Lägg till en notering på en användare";
+    description = "Kicka en person från servern!";
     aliases = [];
     data = new SlashCommandBuilder()
         .setName(this.name)
@@ -22,38 +22,43 @@ export default class NoteCommand implements Command {
         .addUserOption((option) =>
             option
                 .setName("user")
-                .setDescription("Personen du vill lägga till en notering på")
+                .setDescription("Personen du vill kicka")
                 .setRequired(true),
         )
         .addStringOption((option) =>
             option
                 .setName("reason")
-                .setDescription("Anledning till noteringen")
+                .setDescription("Anledning till kicket")
                 .setRequired(true),
         );
     async execute(interaction: ChatInputCommandInteraction) {
         const member = interaction.options.get("user", true)
             .member as GuildMember;
         const reason = interaction.options.get("reason", true).value as string;
-        this.noteUser(member, reason, interaction.user.id);
+        const hasSentMessage = await this.kick(
+            member,
+            reason,
+            interaction.user.id,
+        );
 
-        const noteEmbed = CreateModLogEmbed(
-            "note",
-            "Du har nu laggt en notering på " + member.user.username,
+        const kickEmbed = CreateModLogEmbed(
+            "kick",
+            `${member.user.username} har blivit kickad`,
             reason,
             this.name,
             interaction,
-            true,
+            hasSentMessage,
         );
 
-        await interaction.editReply({ embeds: [noteEmbed] });
+        await interaction.editReply({ embeds: [kickEmbed] });
     }
-    async noteUser(member: GuildMember, reason: string, authorId: string) {
+    async kick(member: GuildMember, reason: string, authorId: string) {
         const userData = await GamerBotAPIInstance.models.getUserData(
             member.user.id,
         );
+
         const modLog = new ModLog(
-            "note",
+            "kick",
             member.user.id,
             member.user.username,
             reason,
@@ -61,8 +66,16 @@ export default class NoteCommand implements Command {
             Date.now(),
             authorId,
         );
-
         userData.modLogs.push(modLog);
         userData.save();
+
+        let hasSentMessage = true;
+        await member
+            .send(
+                `Du har blivit kickad från SGC,\nAnledningen är **${reason}**`,
+            )
+            .catch(() => (hasSentMessage = false));
+        member.kick(reason);
+        return hasSentMessage;
     }
 }
