+++ src/botInteractions/commands/adminCommands/unmute.ts
@@ -7,15 +7,13 @@ import {
 import { Command } from "../../../classes/command.js";
 import { GamerBotAPIInstance } from "../../../index.js";
 import { ModLog } from "../../../classes/modlog.js";
-import { modLogToObject } from "../../../functions/moglogFunctions.js";
-import ms, { StringValue } from "ms";
-import { CreateModLogEmbed } from "../../../functions/builder_functions.js";
+import { CreateModLogEmbed } from "../../../functions/builderFunctions.js";
 
-export default class MuteCommand implements Command {
-    name = "mute";
+export default class UnMuteCommand implements Command {
+    name = "unmute";
     ephemeral = false;
     defer = true;
-    description = "Timar ut personen en viss tid.";
+    description = "Unmutar en användare";
     aliases = [];
     data = new SlashCommandBuilder()
         .setName(this.name)
@@ -24,74 +22,64 @@ export default class MuteCommand implements Command {
         .addUserOption((option) =>
             option
                 .setName("user")
-                .setDescription("Personen du vill muta")
+                .setDescription("Personen du vill unmuta")
                 .setRequired(true),
         )
         .addStringOption((option) =>
             option
                 .setName("reason")
-                .setDescription("Anledning till mutet")
-                .setRequired(true),
-        )
-        .addStringOption((option) =>
-            option
-                .setName("time")
-                .setDescription("Tid du mutar en person")
+                .setDescription("Anledning till unmutet")
                 .setRequired(true),
         );
     async execute(interaction: ChatInputCommandInteraction) {
         const member = interaction.options.get("user", true)
             .member as GuildMember;
         const reason = interaction.options.get("reason", true).value as string;
-        const time =
-            (interaction.options.get("time", false)?.value as string) || "0";
-        const hasSentMessage = await MuteCommand.mute(
+
+        if (!member) return interaction.editReply("Användaren finns inte");
+
+        const hasSentMessage = await this.unMute(
             member,
             reason,
-            time,
             interaction.user.id,
-            `Du har blivit mutead i SGC.\nAnledningen är **${reason}**`,
         );
 
-        const muteEmbed = CreateModLogEmbed(
-            "mute",
-            `${member.user.username} har blivit mutead i ${time}`,
+        const embed = CreateModLogEmbed(
+            "unmute",
+            `${member.user.username} har unmutats`,
             reason,
             this.name,
             interaction,
             hasSentMessage,
         );
 
-        interaction.editReply({ embeds: [muteEmbed] });
+        interaction.editReply({ embeds: [embed] });
     }
-    public static async mute(
-        member: GuildMember,
-        reason: string,
-        time: string,
-        authorId: string,
-        message: string,
-    ) {
+    async unMute(member: GuildMember, reason: string, authorId: string) {
         const userData = await GamerBotAPIInstance.models.getUserData(
             member.id,
         );
-
-        const modLog = new ModLog(
-            "mute",
+        const modlog = new ModLog(
+            "unmute",
             member.id,
             member.user.username,
             reason,
-            time,
+            null,
             Date.now(),
             authorId,
         );
-        userData.modLogs.push(modLogToObject(modLog));
+
+        userData.modLogs.push(modlog);
         userData.save();
 
         let hasSentMessage = true;
-        await member.send(message).catch(() => {
-            hasSentMessage = false;
-        });
-        await member.timeout(ms(time as StringValue), reason);
+
+        await member
+            .send(`Du har blivit unmutad i SGC.\nAnledningen är **${reason}**`)
+            .catch(() => (hasSentMessage = false));
+
+        member.timeout(null);
+
         return hasSentMessage;
     }
 }
