+++ src/botInteractions/commands/adminCommands/userinfo.ts
