+++ src/botInteractions/commands/publicCommand/calc.ts
