+++ src/botInteractions/commands/publicCommand/lvl.ts
