+++ src/botInteractions/commands/publicCommand/ping.ts
