+++ src/botInteractions/commands/publicCommand/serverinfo.ts
