+++ src/botInteractions/commands/publicCommand/uptime.ts
