+++ src/botInteractions/commands/staffCommands/giveframe.ts
