+++ src/botInteractions/messageInteractions/gamerbot.ts
